Backend created....
Created a new Backend class. Hardcoded the setters. Getters are also
included, you should be able to use the index/position number assigned
through ListView in order to ask for a category, or the Uri to display
the image.

I called the constructor on the onCreate method so it creates a database
object in the beginning.  Lets try to use the Backend object as the
single array to populate the lists for Category and Search by word. Let
me know what you guys think