Changed SearchByWord to use Backend class
I modified the SearchByWord activity to do a couple of things.

#1. It now uses the Backend class in order to populate the Array. When
you run the app, the list has everything. Its quicker and dirtier than
Sqlite but it will work.
#2. I implemented an OnItemClickListener, this means that every time you
click an individual item on the list, it is going to toast the file path
that I put into the Uri. I'm not sure how to go from path to picture,
but at the very least we have clickable items that will enable us to do
SOMETHING.