I broke it! But I fixed it!
Okay... So upon further testing I realized my new activity would not
display the correct sign and word if you use the search filter. So what
I did was re coded the whole thing to pass a STRING to DisplaySign and
then created a new method in BackEnd to search through the array and
match the string, and then return position, and then use that to pull up
the picture path etc.