Created a new training game to replace non-working
The current training game would take too much time to debug and rebuild
since it uses SQlite. I short-cutted everything and simplified the setup
so it uses the same image code as my display sign. Instead of creating a
new results screen I created a Toast on the mainscreen if information is
passed to it.

It's quick.. It's dirty.. but it works so far.