Search By Category Complete
Finished search by category. You are able to click a category and/or
picture (i'm not sure how this works on a real phone) And it will bring
up a searchable list of the associated words. It relies heavily on
Jamey's searchable list class, so I heavily modified that by adding some
code in the beginning to see if there is any words being passed through.

Now time to see if we can get anything else to work....