Changed how Results are displayed
Created a popup with the score and asks user if they want to play again.
I used Alertdialog and its built in methods.

I also cleaned up BTGA.java in order to make it more readable and easier
to edit. There are now several new methods to help things be tidy.